Added files
Unfortunately most of the editing was done offline before I was aware we needed to use a git repo to update on changes and progress. I understand the purpose of this (it's both to track progress and check that we're not copying or just cheating) and tried to be as detailed as possible in my journal. I hope that will suffice and any points lost for this error are deserved as it was my mistake